+++ src/main/java/com/project/snsserver/domain/board/repository/jpa/CommentRepository.java
@@ -1,6 +1,7 @@
 package com.project.snsserver.domain.board.repository.jpa;
 
 import com.project.snsserver.domain.board.model.entity.Comment;
+import com.project.snsserver.domain.board.repository.jpa.custom.CustomCommentRepository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
