+++ src/main/java/com/project/snsserver/domain/board/repository/jpa/PostHashtagRepository.java
@@ -1,6 +1,7 @@
 package com.project.snsserver.domain.board.repository.jpa;
 
 import com.project.snsserver.domain.board.model.entity.PostHashtag;
+import com.project.snsserver.domain.board.repository.jpa.custom.CustomPostHashtagRepository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
