+++ src/main/java/com/project/snsserver/domain/board/repository/jpa/PostHeartRepository.java
@@ -2,6 +2,7 @@ package com.project.snsserver.domain.board.repository.jpa;
 
 import com.project.snsserver.domain.board.model.entity.Post;
 import com.project.snsserver.domain.board.model.entity.PostHeart;
+import com.project.snsserver.domain.board.repository.jpa.custom.CustomPostHeartRepository;
 import com.project.snsserver.domain.member.model.entity.Member;
 
 import org.springframework.data.jpa.repository.JpaRepository;
