+++ src/main/java/com/project/snsserver/domain/board/repository/jpa/PostImageRepository.java
@@ -2,6 +2,7 @@ package com.project.snsserver.domain.board.repository.jpa;
 
 import com.project.snsserver.domain.board.model.entity.Post;
 import com.project.snsserver.domain.board.model.entity.PostImage;
+import com.project.snsserver.domain.board.repository.jpa.custom.CustomPostImageRepository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
