+++ src/main/java/com/project/snsserver/domain/board/repository/jpa/PostRepository.java
@@ -1,6 +1,7 @@
 package com.project.snsserver.domain.board.repository.jpa;
 
 import com.project.snsserver.domain.board.model.entity.Post;
+import com.project.snsserver.domain.board.repository.jpa.custom.CustomPostRepository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
