+++ src/main/java/com/project/snsserver/domain/board/repository/jpa/custom/CustomCommentRepository.java
@@ -1,4 +1,4 @@
-package com.project.snsserver.domain.board.repository.jpa;
+package com.project.snsserver.domain.board.repository.jpa.custom;
 
 import com.project.snsserver.domain.board.model.dto.CommentResponse;
