+++ src/main/java/com/project/snsserver/domain/board/repository/jpa/custom/CustomPostHeartRepository.java
@@ -1,4 +1,4 @@
-package com.project.snsserver.domain.board.repository.jpa;
+package com.project.snsserver.domain.board.repository.jpa.custom;
 
 public interface CustomPostHeartRepository {
